modified as per issues

+++ src/VideoMeeting.Application/Features/Meetings/Commands/DisconnectFromMeetingCommand.cs
@@ -4,5 +4,5 @@ namespace VideoMeeting.Application.Features.Meetings.Commands;
 
 public record DisconnectFromMeetingCommand(
     string RoomCode,
-    int UserId
+    long UserId
 ) : IRequest<bool>;
